Cameron Replication with data
on March 26th I downloaded my data from qualtrics, cleaned it by removing all the incomplete responses, and plugged it into the code I made for step H in 4802. The next step is to include the responses that are more than 1/2 done, as well as any new responses that may come in.